Move teleport utils to utils package
